Implementando sistema de pagamento e voucher do pagamento

+++ database/migrations/2018_05_14_202313_create_widget_logs_table.php
@@ -18,7 +18,7 @@ class CreateWidgetLogsTable extends Migration
             $table->timestamps();
             $table->integer('clicks')->default(0);
             $table->integer('impressions')->default(0);
-            $table->integer('revenues')->default(0);
+            $table->double('revenues')->default(0);
         });
 
         Schema::table('widget_logs', function(Blueprint $table) {
